Fixed missing casting
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@8847 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETController.m
@@ -695,7 +695,7 @@ See also ETItemTemplate. */
 	{
 		if (counter > 0)
 		{
-			uniqueKey = [NSString stringWithFormat: @"%@ %lu", key, counter];
+			uniqueKey = [NSString stringWithFormat: @"%@ %lu", key, (unsigned long)counter];
 		}
 		// TODO: Remove the content call once -objectForKey: is included in ETKeyedCollection
 		isUsed = ([[aCollection content] objectForKey: uniqueKey] != nil);
